// src/model/currentWeather.tsx
export interface CurrentWeather {
  dataCollectTime: Date;
  city: string;
  country: string;
  temp_c: number;
  temp_f: number;
  condition: Condition;
  wind_mph: number;
  wind_kph: number;
  wind_dir: string;
  pressure_mb: number;
  pressure_in: number;
  precip_mm: number;
  precip_in: number;
  humidity: number;
  cloud: number;
  feelslike_c: number;
  feelslike_f: number;
  vis_km: number;
  vis_miles: number;
  uv: number;
}

interface Condition {
  text: string;
  icon: string;
}